+++ export/exporter.js
@@ -55,15 +55,15 @@ Exporter.prototype.mapObservations = function(observations) {
 };
 
 Exporter.prototype.requestObservations = function(filter, done) {
-  filter.states = ['active'];
+  const options = {
+    filter: {
+      states: ['active'],
+      observationStartDate: filter.startDate,
+      observationEndDate: filter.endDate
+    }
+  }
 
-  filter.observationStartDate = filter.startDate;
-  delete filter.startDate;
-
-  filter.observationEndDate = filter.endDate;
-  delete filter.endDate;
-
-  new api.Observation(this._event).getAll({filter: filter}, done);
+  new api.Observation(this._event).getAll(options, done);
 };
 
 Exporter.prototype.requestLocations = function(options, done) {