+++ CHANGELOG.md
@@ -6,6 +6,13 @@ MAGE adheres to [Semantic Versioning](http://semver.org/).
 ## Pending on [`develop`](https://github.com/ngageoint/mage-server/tree/develop)
 
 ##### Features
+
+##### Bug Fixes
+
+
+## [5.4.1](https://github.com/ngageoint/mage-server/releases/tag/5.4.1)
+
+##### Features
 * Added observation location provider and location accuracy exports.
 
 ##### Bug Fixes